still issues on zoom when particles are in small area, zoom becomes >0 which make display upside down
git-svn-id: http://svn.lam.fr/repos/glnemo2/trunk@316 6a5dd85c-8a21-4469-84b3-770a23de0f29

+++ src/globaloptions.h
@@ -62,7 +62,7 @@ public:
     bool cam_display_spline;
     bool cam_play;
     // from Scene Orientation TAB
-    float zoom,zoomo;
+    double zoom,zoomo;
     float xrot,yrot,zrot;
     float urot,vrot,wrot;
     float ixrot,iyrot,izrot;
